fix(core): Fixing core user offline state

+++ actor-apps/core/src/main/java/im/actor/core/modules/updates/PresenceProcessor.java
@@ -31,7 +31,7 @@ public class PresenceProcessor extends AbsModule {
     }
 
     @Verified
-    public void onUserLastSeen(int uid, long date, long updateDate) {
+    public void onUserLastSeen(int uid, int date, long updateDate) {
         presenceActor.sendOnce(new PresenceActor.UserLastSeen(uid, date, updateDate));
     }
